feat(ui): inject secrets - WF-428
