+++ src/ui/src/builder/BuilderVault.vue
@@ -4,58 +4,9 @@ import { useKeyValueEditor } from "./settings/composables/useKeyValueEditor";
 import BuilderTemplateInput from "./settings/BuilderTemplateInput.vue";
 import WdsFieldWrapper from "@/wds/WdsFieldWrapper.vue";
 import WdsButton from "@/wds/WdsButton.vue";
-import { computed, onMounted, ref, shallowRef } from "vue";
+import { inject, onMounted } from "vue";
 import WdsSkeletonLoader from "@/wds/WdsSkeletonLoader.vue";
-import { JSONValue } from "./settings/BuilderFieldsKeyValue.vue";
-
-function useSecretsManager() {
-	const secrets = shallowRef<JSONValue>({});
-	const isLoading = ref(false);
-	const isSaving = ref(false);
-	const errror = ref(false);
-
-	async function load() {
-		isLoading.value = true;
-		errror.value = false;
-
-		try {
-			await new Promise((res) => setTimeout(res, 500));
-			// TODO: real implementation
-			secrets.value = {
-				GOOGLE_SECRET: "foo",
-				WRITER_API_KEY: "foo",
-			};
-		} catch (e) {
-			errror.value = e;
-			secrets.value = {};
-		} finally {
-			isLoading.value = false;
-		}
-	}
-
-	async function update(newSecrets: JSONValue) {
-		isSaving.value = true;
-		errror.value = false;
-
-		try {
-			await new Promise((res) => setTimeout(res, 500));
-			// TODO: real implementation
-			secrets.value = newSecrets;
-		} catch (e) {
-			errror.value = e;
-		} finally {
-			isSaving.value = false;
-		}
-	}
-
-	return {
-		secrets,
-		load,
-		readonly: computed(() => isLoading.value || isSaving.value),
-		isLoading,
-		update,
-	};
-}
+import injectionKeys from "@/injectionKeys";
 
 const {
 	secrets,
@@ -63,7 +14,7 @@ const {
 	readonly,
 	isLoading: isSecretsLoading,
 	update: updateSecrets,
-} = useSecretsManager();
+} = inject(injectionKeys.secretsManager);
 
 const {
 	currentValue,
