+++ src/ui/src/builder/settings/BuilderTemplateInput.vue
@@ -79,6 +79,7 @@ import Fuse from "fuse.js";
 import injectionKeys from "@/injectionKeys";
 import {
 	PropType,
+	computed,
 	inject,
 	nextTick,
 	onUnmounted,
@@ -89,6 +90,13 @@ import {
 import WdsTextInput from "@/wds/WdsTextInput.vue";
 import WdsTextareaInput from "@/wds/WdsTextareaInput.vue";
 import { useFloating, size, flip, autoUpdate } from "@floating-ui/vue";
+
+const {
+	secrets,
+	load: loadSecrets,
+	isLoading: isSecretsLoading,
+	update: updateSecrets,
+} = inject(injectionKeys.secretsManager);
 
 const emit = defineEmits(["input", "update:value"]);
 
@@ -222,6 +230,16 @@ function handleInput(ev) {
 	showAutocomplete();
 }
 
+const autoCompletionState = computed(() => {
+	const userState = ss.userState.value ?? {};
+	if (!secrets.value) return userState;
+
+	return {
+		...userState,
+		vault: secrets.value,
+	};
+});
+
 function showAutocomplete() {
 	const { selectionStart, selectionEnd } = input.value?.getSelection() ?? {};
 	const newValue = input.value?.value;
@@ -239,12 +257,12 @@ function showAutocomplete() {
 	const keyword = full.at(-1);
 	const path = full.slice(0, -1);
 
-	const allOptions = Object.entries(_get(ss.userState.value, path) ?? {}).map(
-		([key, val]) => ({
-			text: escapeVariable(key),
-			type: typeToString(val),
-		}),
-	);
+	const allOptions = Object.entries(
+		_get(autoCompletionState.value, path) ?? {},
+	).map(([key, val]) => ({
+		text: escapeVariable(key),
+		type: typeToString(val),
+	}));
 
 	const fuse = new Fuse(allOptions, {
 		findAllMatches: true,
