+++ src/ui/src/writerTypes.ts
@@ -4,6 +4,7 @@ import { generateBuilderManager, Selection } from "./builder/builderManager";
 import type { SchemaObject } from "ajv";
 import type { useNotesManager } from "./core/useNotesManager";
 import type { useCollaborationManager } from "./composables/useCollaborationManager";
+import type { useSecretsManager } from "./core/useSecretsManager";
 
 export type Core = ReturnType<typeof generateCore>;
 
@@ -70,9 +71,9 @@ export type WriterComponentDefinitionField = {
 	/** Which control (text, textarea, etc) to use if not the default for the type */
 	control?: FieldControl;
 	options?:
-	| Record<string, string>
-	| string // For predefined functions
-	| ((wf?: Core, componentId?: ComponentId) => Record<string, string>); // List of values to be provided as autocomplete options
+		| Record<string, string>
+		| string // For predefined functions
+		| ((wf?: Core, componentId?: ComponentId) => Record<string, string>); // List of values to be provided as autocomplete options
 	/** Data type for the field */
 	type: FieldType;
 	/** Category (Layout, Content, etc) */
@@ -129,6 +130,8 @@ export type WriterComponentDefinition = {
 export type BuilderManager = ReturnType<typeof generateBuilderManager>;
 
 export type NotesManager = ReturnType<typeof useNotesManager>;
+
+export type SecretsManager = ReturnType<typeof useSecretsManager>;
 
 export type CollaborationManager = ReturnType<typeof useCollaborationManager>;
